% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{My.plot.xmean.ordinaly}
\alias{My.plot.xmean.ordinaly}
\title{My plot.xmean.ordinaly from rms
removed the xlab and ylab defaults from plot.xmean.ordinaly - so I could set my own}
\usage{
My.plot.xmean.ordinaly(x, data, subset, na.action, subn = TRUE,
  cr = FALSE, topcats = 1, cex.points = 0.75, ...)
}
\description{
My plot.xmean.ordinaly from rms
removed the xlab and ylab defaults from plot.xmean.ordinaly - so I could set my own
}
