% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{My.val.prob.ci.2}
\alias{My.val.prob.ci.2}
\title{my val.prb ci.2 function
My hack of steyerberg's cal.plot function - which is an adapataion of the val.prob function from rms
The main thing was to get all hist bins poining up to match plot(calibrate) - I had 2 panel cal plot - needed to match the bottom histograms
couldn't do that with val.prob because no way to control space under hist plot - could do that with this function
changed degree to 1 to match val.prob which is what I really wanted to use in the first place
how to install the package
library(githubinstall)
githubinstall("CalibrationCurves")
library(CalibrationCurves)}
\usage{
My.val.prob.ci.2(p, y, logit, group, weights = rep(1, length(y)),
  normwt = F, pl = T, smooth = c("loess", "rcs", F),
  CL.smooth = "fill", CL.BT = F, lty.smooth = 1,
  col.smooth = "black", lwd.smooth = 1, nr.knots = 5,
  logistic.cal = F, lty.log = 1, col.log = "black", lwd.log = 1,
  xlab = "Predicted Probability", ylab = "Observed proportion",
  xlim = c(-0.02, 1), ylim = c(-0.15, 1), m, g, cuts, emax.lim = c(0,
  1), legendloc = c(0.5, 0.27), statloc = c(0, 0.85), dostats = T,
  cl.level = 0.95, method.ci = "pepe", roundstats = 2,
  riskdist = "predicted", cex = 0.75, cex.leg = 0.75,
  connect.group = F, connect.smooth = T, g.group = 4,
  evaluate = 100, nmin = 0, d0lab = "0", d1lab = "1",
  cex.d01 = 0.7, dist.label = 0.04, line.bins = -0.05,
  dist.label2 = 0.03, cutoff, las = 1, length.seg = 1,
  y.intersp = 1, lty.ideal = 1, col.ideal = "grey", lwd.ideal = 1,
  ...)
}
\description{
my val.prb ci.2 function
My hack of steyerberg's cal.plot function - which is an adapataion of the val.prob function from rms
The main thing was to get all hist bins poining up to match plot(calibrate) - I had 2 panel cal plot - needed to match the bottom histograms
couldn't do that with val.prob because no way to control space under hist plot - could do that with this function
changed degree to 1 to match val.prob which is what I really wanted to use in the first place
how to install the package
library(githubinstall)
githubinstall("CalibrationCurves")
library(CalibrationCurves)
}
\examples{
My.val.prob.ci.2()
}
\keyword{Calibration}
\keyword{curves}
