% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MyValPlot}
\alias{MyValPlot}
\title{MyValPlot function
My hack val.prob for my plot - set line type and legend etc}
\usage{
MyValPlot(x, xlab, ylab, xlim, ylim, legend = TRUE, subtitles = TRUE,
  scat1d.opts = NULL, ...)
}
\description{
MyValPlot function
My hack val.prob for my plot - set line type and legend etc
}
\examples{
MyValPlot()
}
\keyword{Validation}
\keyword{plot}
